move metrics editor related files to a separate folder
