+++ public/app/plugins/datasource/cloudwatch/components/MetricsQueryEditor/MetricsQueryHeader.tsx
@@ -4,8 +4,8 @@ import { SelectableValue } from '@grafana/data';
 import { InlineSelect, FlexItem } from '@grafana/experimental';
 import { Button, ConfirmModal, RadioButtonGroup } from '@grafana/ui';
 
-import { CloudWatchDatasource } from '../datasource';
-import { CloudWatchMetricsQuery, CloudWatchQuery, MetricEditorMode, MetricQueryType } from '../types';
+import { CloudWatchDatasource } from '../../datasource';
+import { CloudWatchMetricsQuery, CloudWatchQuery, MetricEditorMode, MetricQueryType } from '../../types';
 
 interface MetricsQueryHeaderProps {
   query: CloudWatchMetricsQuery;
